astro/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pymysql
boto3
pandas
apache-airflow-providers-apache-spark
apache-airflow-providers-microsoft-mssql
sqlalchemy
pyodbc
python-dotenv
pyspark==3.5.3
delta-spark==3.2.0
apache-airflow-providers-mysql
apache-airflow-providers-amazon
mysqlclient
